requirements.txt
pygame==2.6.1
supabase>=2.4.0
python-dotenv==1.0.0 